// src/types/generated/ProjectView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FolderPrefs } from "./FolderPrefs";
import type { SourceFile } from "./SourceFile";
import type { WorldType } from "./WorldType";

/**
 * Result of opening a folder (O6/O7/O8 bundled).
 */
export type ProjectView = { folder: string, files: Array<SourceFile>, total_dialogue_lines: number, detected_source_lang: string | null, detected_world: WorldType, prefs: FolderPrefs, supports_glossary: boolean, };